+++ backend/tests/external_dependency_unit/conftest.py
@@ -1,0 +1,61 @@
+from collections.abc import Generator
+from uuid import uuid4
+
+import pytest
+from fastapi_users.password import PasswordHelper
+from sqlalchemy.orm import Session
+
+from onyx.db.engine.sql_engine import get_session_with_current_tenant
+from onyx.db.engine.sql_engine import SqlEngine
+from onyx.db.models import User
+from onyx.db.models import UserRole
+from shared_configs.contextvars import CURRENT_TENANT_ID_CONTEXTVAR
+from tests.external_dependency_unit.constants import TEST_TENANT_ID
+
+
+@pytest.fixture(scope="function")
+def db_session() -> Generator[Session, None, None]:
+    """Create a database session for testing using the actual PostgreSQL database"""
+    # Make sure that the db engine is initialized before any tests are run
+    SqlEngine.init_engine(
+        pool_size=10,
+        max_overflow=5,
+    )
+    with get_session_with_current_tenant() as session:
+        yield session
+
+
+@pytest.fixture(scope="function")
+def tenant_context() -> Generator[None, None, None]:
+    """Set up tenant context for testing"""
+    # Set the tenant context for the test
+    token = CURRENT_TENANT_ID_CONTEXTVAR.set(TEST_TENANT_ID)
+    try:
+        yield
+    finally:
+        # Reset the tenant context after the test
+        CURRENT_TENANT_ID_CONTEXTVAR.reset(token)
+
+
+def create_test_user(db_session: Session, email_prefix: str) -> User:
+    """Helper to create a test user with a unique email"""
+    # Use UUID to ensure unique email addresses
+    unique_email = f"{email_prefix}_{uuid4().hex[:8]}@example.com"
+
+    password_helper = PasswordHelper()
+    password = password_helper.generate()
+    hashed_password = password_helper.hash(password)
+
+    user = User(
+        id=uuid4(),
+        email=unique_email,
+        hashed_password=hashed_password,
+        is_active=True,
+        is_superuser=False,
+        is_verified=True,
+        role=UserRole.EXT_PERM_USER,
+    )
+    db_session.add(user)
+    db_session.commit()
+    db_session.refresh(user)
+    return user
