+++ backend/tests/external_dependency_unit/constants.py
@@ -1,0 +1,1 @@
+TEST_TENANT_ID: str = "public"
