+++ backend/tests/external_dependency_unit/file_store/test_file_store_non_mocked.py
@@ -19,17 +19,16 @@ from sqlalchemy.orm import Session
 
 from onyx.configs.constants import FileOrigin
 from onyx.db.engine.sql_engine import get_session_with_current_tenant
-from onyx.db.engine.sql_engine import SqlEngine
 from onyx.file_store.file_store import S3BackedFileStore
 from onyx.utils.logger import setup_logger
 from shared_configs.contextvars import CURRENT_TENANT_ID_CONTEXTVAR
+from tests.external_dependency_unit.constants import TEST_TENANT_ID
 
 logger = setup_logger()
 
 
 TEST_BUCKET_NAME: str = "onyx-file-store-tests"
 TEST_FILE_PREFIX: str = "test-files"
-TEST_TENANT_ID: str = "public"
 
 
 # Type definitions for test data
@@ -102,30 +101,6 @@ def _get_all_backend_configs() -> List[BackendConfig]:
         )
 
     return configs
-
-
-@pytest.fixture(scope="function")
-def db_session() -> Generator[Session, None, None]:
-    """Create a database session for testing using the actual PostgreSQL database"""
-    # Make sure that the db engine is initialized before any tests are run
-    SqlEngine.init_engine(
-        pool_size=10,
-        max_overflow=5,
-    )
-    with get_session_with_current_tenant() as session:
-        yield session
-
-
-@pytest.fixture(scope="function")
-def tenant_context() -> Generator[None, None, None]:
-    """Set up tenant context for testing"""
-    # Set the tenant context for the test
-    token = CURRENT_TENANT_ID_CONTEXTVAR.set(TEST_TENANT_ID)
-    try:
-        yield
-    finally:
-        # Reset the tenant context after the test
-        CURRENT_TENANT_ID_CONTEXTVAR.reset(token)
 
 
 @pytest.fixture(
